+++ packages/fighting-design/components.ts
@@ -126,6 +126,5 @@ export type { RadioGroupInstance } from './radio-group'
 export { FImagePreview } from './image-preview'
 export type { FImagePreviewInstance } from './image-preview'
 
-export { FExpandCard } from './expandcard'
-export type { FExpandCardInstance } from './expandcard'
-
+export { FExpandCard } from './expand-card'
+export type { FExpandCardInstance } from './expand-card'