+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.6.5-alpha.1 (2022-09-16)
+
+- 新增 `f-expand-card` 组件
 - 优化 `f-toolbar-item` 和 `f-toolbar` 组件样式细节
 - `f-toolbar-item` 新增 `data-key` `icon` `icon-size` 配置项
 - `f-toolbar` 新增 `click` 回调参数
-- 新增 `f-expand-card` 组件
 - `f-toolbar` 针对于苹果 `Safari` 点击失效做出兼容
 - `f-toolbar` 组件废除 `vague` 配置项